add ripu mining algorithm
