+++ ymir/start.py
@@ -10,14 +10,14 @@ def main() -> int:
     gpu_id: str = str(ymir_cfg.param.get('gpu_id', '0'))
     gpu_count: int = ymir_cfg.param.get('gpu_count', None) or len(gpu_id.split(','))
 
+    mining_cmd = 'python3 ymir/ymir_mining.py'
     if gpu_count <= 1:
         infer_cmd = 'python3 ymir/ymir_infer.py'
-        mining_cmd = 'python3 ymir/ymir_mining.py'
     else:
         port = find_free_port()
         dist_cmd = f'-m torch.distributed.launch --master_port {port} --nproc_per_node {gpu_count}'
         infer_cmd = f'python3 {dist_cmd} ymir/ymir_infer.py'
-        mining_cmd = f'python3 {dist_cmd} ymir/ymir_mining.py'
+        
     apps = dict(training='python3 ymir/ymir_training.py', mining=mining_cmd, infer=infer_cmd)
     executor = Executor(apps)
     executor.start()
